move useFilteredIndices / useCategoryFilterIndices to hooks.ts
