+++ src/react/scatter_state.ts
@@ -1,8 +1,8 @@
 import { Matrix4 } from '@math.gl/core';
 import type { PickingInfo } from "deck.gl/typed";
-import type { ScatterPlotConfig, VivRoiConfig } from "./components/VivMDVReact";
-import { useChart, useDataStore } from "./context";
-import { useChartID, useChartSize, useConfig, useParamColumns } from "./hooks";
+import type { ScatterPlotConfig } from "./components/VivMDVReact";
+import { useChart } from "./context";
+import { useChartID, useChartSize, useConfig, useFilteredIndices, useParamColumns } from "./hooks";
 import { useCallback, useEffect, useMemo, useRef, useState } from "react";
 import { getVivId } from "./components/avivatorish/MDVivViewer";
 import { useMetadata } from "./components/avivatorish/state";
@@ -11,76 +11,9 @@ import type { ViewState } from './components/VivScatterComponent';
 import SpatialLayer from '@/webgl/SpatialLayer';
 import { ScatterSquareExtension, ScatterDensityExension } from '../webgl/ScatterDeckExtension';
 import { useHighlightedIndex } from './selectionHooks';
-import type { DataColumn } from '@/charts/charts';
 import { useLegacyDualContour } from './contour_state';
 
-/**
- * Get a {Uint32Array} of the currently filtered indices.
- * When the selection changes, this will asynchronously update.
- * All users of the same data store (on a per-chart basis) share a reference to the same array.
- * -- change properties/settings so that we can more dynamically select.
- */
-export function useFilteredIndices() {
-    // in the case of region data, it should be filtered by that as well...
-    // I really want to sort out how I use types here...
-    const config = useConfig<VivRoiConfig>();
-    const filterColumn = config.background_filter?.column;
-    const dataStore = useDataStore();
-    const [filteredIndices, setFilteredIndices] = useState(new Uint32Array());
-    const [filteredOutIndices, setFilteredOutIndices] = useState(new Uint32Array());
-    // biome-ignore lint/correctness/useExhaustiveDependencies: shouldn't be ignoring this, but some deps don't tend to change as of now.
-    useEffect(() => {
-        // return
-        let cancelled = false;
-        if (!filterColumn) return;
-        const indexPromise = dataStore.getFilteredIndices();
-        //todo maybe make more use of deck.gl category filters once we update to new version
-        const catFilters = [config.background_filter, ...config.category_filters.filter(f => f.category !== 'all')];
-        const catColumns = catFilters.map(f => f.column);
-        const colPromise = window.mdv.chartManager?._getColumnsAsync(dataStore.name, catColumns);
-        Promise.all([indexPromise, colPromise]).then(([indices]) => {
-            if (cancelled) return;
-            if (filterColumn) {
-                const cols = catFilters.map(({column}) => dataStore.columnIndex[column]);
-                const filterValue = config.background_filter?.category;
-                if (filterValue) {
-                    //const filterIndex = col.values.indexOf(filterValue);
-                    const filterIndex = catFilters.map(f => {
-                        if (Array.isArray(f.category)) return f.category.map(c => dataStore.columnIndex[f.column].values.indexOf(c)) as number[];
-                        return dataStore.columnIndex[f.column].values.indexOf(f.category) as number;
-                    });
-                    try {
-                        // const filteredIndices = indices.filter(i => col.data[i] === filterIndex);
-                        const filteredIndices = indices.filter(i => catFilters.every((_, j) => {
-                            const f = filterIndex[j];
-                            if (typeof f === 'number') return f === cols[j].data[i];
-                            return f.some(fi => cols[j].data[i] === fi);
-                        }));
-                        setFilteredIndices(filteredIndices);
-                        // thinking about allowing gray-out of non-selected points... should be optional
-                        // const filteredOutIndices = indices.filter(i => col.data[i] !== filterIndex);
-                        // setFilteredOutIndices(filteredOutIndices);
-                    } catch (e) {
-                        console.error('error filtering indices', e);
-                        return;
-                    }
-                    return;
-                }
-            }
-            setFilteredIndices(indices);
-        });
-        // should I have a cleanup function to cancel the promise if it's not resolved
-        // by the time the effect is triggered again?
-        return () => {
-            // if (!finished) console.log('filtered indices promise cancelled');
-            cancelled = true;
-        }
 
-        // using _filteredIndicesPromise as a dependency is working reasonably well,
-        // but possibly needs a bit more thought.
-    }, [dataStore._filteredIndicesPromise, filterColumn, config.background_filter, config.category_filters]);
-    return filteredIndices;
-}
 
 export function useRegionScale() {
     const metadata = useMetadata();
